checker/requirements.txt
git+https://github.com/domenukk/enochecker.git@benni
asyncio
pycryptodome
numpy
hypercorn
construct